@@ -2,8 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App;
+use App\Http\Controllers\Controller;
 
 class HomeController extends Controller{
 
