@@ -5,7 +5,10 @@
 @section('content')
 <h1>Обьявления</h1>
 
-<ul>
+<p>
+  <a href="ads/create" class="btn btn-success btn-lg">Создать обьявление</a>
+</p>
+
   @foreach($ads as $ad)
     <div class="jumbotron">
       <h1 class="display-4">{{ $ad->title }}</h1>
@@ -14,5 +17,4 @@
       </p>
     </div>
   @endforeach
-</ul>
 @endsection
